backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
qdrant-client==1.16.1
python-dotenv==1.0.0
sentence-transformers>=2.3.0
torch==2.2.0
pydantic==2.5.0
numpy<2
cohere>=5.0.0
